// src/models/categories.ts
export enum Categories {
  Monitor = " مانیتور",
  Tablet = " تبلت ",
  PrinterAndScanner = " پرینتر-اسکنر|ماشینهای اداری",
  LaserPrinter = " پرینتر لیزری تک کاره|ماشینهای اداری ",
  MultiJobPrinter = " پرینتر چندکاره|ماشینهای اداری ",
  InkjetPrinter = " پرینتر جوهرافشان|ماشینهای اداری ",
  DotMatrixPrinter = " پرینتر سوزنی|ماشینهای اداری ",
  Scanner = " اسکنر|ماشینهای اداری ",
  InternalSSD = " حافظه SSD اینترنال|کامپیوتر ",
  PcParts = " قطعات اصلی|کامپیوتر",
  ExternalSSD = " حافظه SSD اکسترنال|کامپیوتر ",
  InternalHard = " هارد اینترنال",
  ExternalHard = " هارد اکسترنال",
  PrinterLabel = " پرینتر برچسب-بارکد|ماشینهای اداری ",
  PrinterLabel2 = " لیبل|لوازم جانبی ",
  PrinterEssentials = " ملزومات پرینتر",
  KeyBoard = " کیبورد|لوازم جانبی|لوازم جانبی ",
  Speaker = " اسپیکر|لوازم جانبی|لوازم جانبی ",
  ServerRam = " رم سرور|سرور و ملزومات|سرور و ملزومات ",
  Ram = " رم RAM|کامپیوتر ",
  Cpu = " پردازنده CPU|کامپیوتر ",
  Accessories = " لوازم جانبی کامپیوتر ",
  Microphone = " میکروفن",
  Mouse = " ماوس",
  PowerBank = " پاور بانک|لوازم جانبی|لوازم جانبی ",
  Power = " منبع تغذیه Power",
  MainParts = " قطعات اصلی ",
  Laptop = "لپ تاپ|لپ تاپ ",
  MousePad = " پد ماوس|لوازم جانبی|لوازم جانبی ",
  Motherboard = " مادربرد",
  Phone = " موبایل|موبایل و تبلت",
  Mobile = " موبایل",
  MobileAndTablet = "موبایل و تبلت",
  Hub = " هاب USB",
  HandsFree = " هندزفری بی‌سیم",
  Headphone = " هدفون، هدست",
  Convertor = " کابل و مبدل|لوازم جانبی|لوازم جانبی ",
  FlashMemory = " فلش مموری|لوازم جانبی|لوازم جانبی ",
  Projector = " پروژکتور|ماشینهای اداری",
  VoiceCard = " کارت صدا|کامپیوتر",
  VoiceCard1 = " کارت صدا|کامپیوتر",
  MiniPC = " مینی پیسی",
  Case = " کیس CASE",
  AllInOne = " آل‌این‌وان All in one|کامپیوتر|کامپیوتر ",
  ProcessorFan = " فن پردازنده|کامپیوتر ",
  ServerHard = " هارد سرور",
  GraphicCard = " کارت گرافیک VGA|کامپیوتر",
  AnalogCCTV = " دوربین مداربسته AHD|دوربین مداربسته ",
  networkCCTV = " دوربین مداربسته IP",
}